fix(resource): allow modDocument to msProduct type conversion on update
Product Update processor failed to load the resource when class_key was
changed in settings because parent initialize() requires msProduct in DB.
Extract EnsureTargetClassKeyTrait from Category/Update and reuse in
Product/Update. Closes #329.
