+++ core/components/minishop3/src/Processors/Category/Update.php
@@ -2,39 +2,19 @@
 
 namespace MiniShop3\Processors\Category;
 
-use MODX\Revolution\Processors\Resource\Update as UpdateProcessor;
 use MiniShop3\Model\msCategory;
-use MODX\Revolution\modResource;
+use MiniShop3\Processors\Resource\EnsureTargetClassKeyTrait;
+use MODX\Revolution\Processors\Resource\Update as UpdateProcessor;
 
 class Update extends UpdateProcessor
 {
+    use EnsureTargetClassKeyTrait;
+
     public $classKey = msCategory::class;
     public $languageTopics = ['resource', 'minishop3:default'];
     public $permission = 'mscategory_save';
     public $beforeSaveEvent = 'OnBeforeDocFormSave';
     public $afterSaveEvent = 'OnDocFormSave';
-
-    /**
-     * @return bool|null|string
-     */
-    public function initialize()
-    {
-        $primaryKey = $this->getProperty($this->primaryKeyField, false);
-        if (empty($primaryKey)) {
-            return $this->modx->lexicon($this->classKey . '_err_ns');
-        }
-
-        if (!$this->modx->getCount($this->classKey, ['id' => $primaryKey, 'class_key' => $this->classKey])) {
-            $res = $this->modx->getObject(modResource::class, ['id' => $primaryKey]);
-            if ($res) {
-                $res->set('class_key', $this->classKey);
-                $res->save();
-            }
-        }
-
-        return parent::initialize();
-    }
-
 
     /**
      * @return int|mixed|string
